+++ csvEditorHtml/browser/beforeDomLoadedBrowser.ts
@@ -34,6 +34,7 @@ var initialConfig: EditCsvConfig | undefined = {
 	writeOption_escapeChar: '"',
 	writeOption_hasHeader: "false",
 	doubleClickColumnHandleForcedWith: 200,
+	doubleClickRowHandleForcedHeight: 106,
 	openSourceFileAfterApply: false,
 	selectTextAfterBeginEditCell: false,
 	quoteAllFields: false,