+++ CHANGELOG.md
@@ -2,13 +2,13 @@
 
 ## [Unreleased]
 
-# 0.7.7
+# 0.8.0
 - trim cells can now be undone (not for header cells)
 - fixed issue #122 - add option customize text color
 - added feature #130 - add feature resize rows
 	- works the same as `doubleClickColumnHandleForcedWith` (auto resize is a bit better implemented)
 - adds feature #109 - open urls in browser
-	- controled via `convertUrlsToTags` setting
+	- controled via `convertUrlsToLinkTags` setting
 
 # 0.7.6
 - fixed issue #110 - Add feature to swap rows and columns